Correlation Analysis
Correlations between attributes were analysed for each participant individually as well as for the whole data set. In case of individual analysis some correlation are visible on the heatmap, however they are not consistent during comparison of different individual analysis. In case of correlations on the whole data set, no meaningfull correlations were identified.